Renamed the user redux folder since now we're going to store all users, not only the current user.
